requirements/windows.txt
# Windows Optimized Requirements (Stable)
# Use: pip install -r requirements-win.txt

# Core UI & Utils
streamlit==1.50.0
streamlit-pdf-viewer==0.0.26
pymupdf==1.26.5
pymupdf4llm==0.2.9
python-dotenv==1.1.1
pyyaml==6.0.3
psutil
nest-asyncio==1.6.0

# PyTorch - Explicitly targeting stable CPU/CUDA builds
# Note: torchvision is EXCLUDED to prevent 0xc0000139 DLL conflicts
torch==2.5.1 --index-url https://download.pytorch.org/whl/cpu
# For CUDA 12.1 (Uncomment below and comment above if you have a GPU)
# torch==2.5.1 --index-url https://download.pytorch.org/whl/cu121

# RAG & Embeddings
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0
faiss-cpu>=1.9.0
rank-bm25>=0.2.2
optimum[onnxruntime]==2.1.0
onnxruntime==1.19.2

# API Server
fastapi==0.115.10
uvicorn==0.30.6
python-multipart==0.0.22

# LLM Client
ollama==0.6.0

# Testing
pytest==8.3.2
pytest-mock==3.14.0
pytest-cov==5.0.0
pytest-asyncio==0.24.0
